+++ types.go
@@ -1,19 +1,13 @@
 package faktory_worker
 
-import (
-	faktory "github.com/contribsys/faktory/client"
-)
+import "context"
 
 const (
-	Version = "1.3.1"
+	Version = "1.4.0"
 )
 
 // Perform actually executes the job.
 // It must be thread-safe.
-type Perform func(ctx Context, args ...interface{}) error
+type Perform func(ctx context.Context, args ...interface{}) error
 
-type Handler func(ctx Context, job *faktory.Job) error
 type LifecycleEventHandler func(*Manager) error
-
-// MiddlewareFunc defines a function to process middleware.
-type MiddlewareFunc func(Handler) Handler